refactor(App.tsx): extract sidebar JSX + N-1 fetch + highlight pipeline
Options 1 + 2 from the App.tsx size-reduction plan. App.tsx shrinks from
1575 lines to ~1150 by relocating pure presentational JSX and two
self-contained effect pipelines. No behavior change.

Extracted components (Option 1):
  - components/SidebarSummary.tsx  — sticky contingency/N-1 strip
  - components/AppSidebar.tsx      — sidebar layout shell with
                                     summary + contingency selector
                                     + children slot
  - components/StatusToasts.tsx    — error/info floating banners

Extracted hooks (Option 2):
  - hooks/useN1Fetch.ts            — N-1 diagram fetch effect:
                                     svgPatch fast-path, full
                                     /api/n1-diagram fallback, and
                                     contingency-change confirm
                                     routing
  - hooks/useDiagramHighlights.ts  — per-tab SVG highlight pipeline
                                     (overload halos, contingency
                                     highlight, action-target halos,
                                     delta visuals) + the per-tab
                                     Flow/Impacts view-mode state

One unavoidable eslint-disable in useDiagramHighlights.ts on the
reattach-prune setState: the pre-extraction code in App.tsx was not
flagged by react-hooks/set-state-in-effect (position-sensitive
heuristic), but the observable behavior must be preserved byte-for-
byte, so the guarded setState stays with a justified suppression.

frontend/CLAUDE.md records what landed, lists deferred options
(Option 3 orchestrator hooks, Option 4 AppContext, the NDJSON parser
extraction), and explains the ESLint exception.

Verified: npm run test (1098 tests pass), npm run lint clean, tsc -b
clean, scripts/check_standalone_parity.py OK,
scripts/check_code_quality.py OK.
