+++ frontend/src/App.tsx
@@ -5,20 +5,19 @@
 // SPDX-License-Identifier: MPL-2.0
 // This file is part of Co-Study4Grid a Power Grid Study tool Assistant Interface to help solve contigencies for a grid state under study. 
 
-import React, { useState, useEffect, useCallback, useRef, useMemo } from 'react';
+import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
 import './App.css';
 import VisualizationPanel from './components/VisualizationPanel';
 import ActionFeed from './components/ActionFeed';
 import OverloadPanel from './components/OverloadPanel';
 import Header from './components/Header';
+import AppSidebar from './components/AppSidebar';
+import StatusToasts from './components/StatusToasts';
 import SettingsModal from './components/modals/SettingsModal';
 import ReloadSessionModal from './components/modals/ReloadSessionModal';
 import ConfirmationDialog from './components/modals/ConfirmationDialog';
 import type { ConfirmDialogState } from './components/modals/ConfirmationDialog';
 import { api } from './api';
-import { applyOverloadedHighlights, applyDeltaVisuals, applyActionTargetHighlights, applyContingencyHighlight, processSvg } from './utils/svgUtils';
-import { cloneBaseSvg, applyPatchToClone } from './utils/svgPatch';
-import { computeN1OverloadHighlights } from './utils/overloadHighlights';
 import type { ActionDetail, ActionOverviewFilters, DiagramData, TabId, RecommenderDisplayConfig, UnsimulatedActionScoreInfo } from './types';
 import { useSettings } from './hooks/useSettings';
 import { useActions } from './hooks/useActions';
@@ -27,6 +26,8 @@ import { useDiagrams } from './hooks/useDiagrams';
 import { useSession } from './hooks/useSession';
 import { useDetachedTabs } from './hooks/useDetachedTabs';
 import { useTiedTabsSync, type PZInstance } from './hooks/useTiedTabsSync';
+import { useN1Fetch } from './hooks/useN1Fetch';
+import { useDiagramHighlights } from './hooks/useDiagramHighlights';
 import { interactionLogger } from './utils/interactionLogger';
 
 function App() {
@@ -878,118 +879,18 @@ function App() {
 
 
 
-  useEffect(() => {
-    if (!selectedBranch) {
-      diagrams.setN1Diagram(null);
-      if (!hasAnalysisState()) {
-        diagrams.committedBranchRef.current = '';
-      }
-      return;
-    }
-
-    if (branches.length > 0 && !branches.includes(selectedBranch)) return;
-
-    // Short-circuit when we already have the N-1 diagram / fetch /
-    // analysis for this branch — EXCEPT on session restore, where
-    // `hasAnalysisState()` is already true (actions were rehydrated
-    // by `useSession::handleRestoreSession`) while `n1Diagram` is
-    // still null. In that case the early-return would silently skip
-    // the N-1 fetch, leaving the N-1 tab blank and the N-1 overloads
-    // panel empty. `restoringSessionRef.current` is the signal we
-    // use to force the fetch through.
-    if (selectedBranch === diagrams.committedBranchRef.current
-        && !diagrams.restoringSessionRef.current
-        && (n1Diagram || hasAnalysisState() || n1Loading || analysisLoading)) return;
-
-    if (selectedBranch !== diagrams.committedBranchRef.current && hasAnalysisState() && !diagrams.restoringSessionRef.current) {
-      setConfirmDialog({ type: 'contingency', pendingBranch: selectedBranch });
-      setSelectedBranch(diagrams.committedBranchRef.current);
-      return;
-    }
-    // Capture BEFORE we reset the ref so the rest of this effect can
-    // distinguish "user changed contingency" from "session was just
-    // restored". Without the local capture, any later branch in
-    // this effect would see `restoringSessionRef.current === false`
-    // and incorrectly wipe the just-restored analysis state via
-    // `clearContingencyState()`.
-    const isRestoring = diagrams.restoringSessionRef.current;
-    diagrams.restoringSessionRef.current = false;
-
-    diagrams.committedBranchRef.current = selectedBranch;
-    if (!isRestoring) {
-      clearContingencyState();
-      diagrams.setN1Diagram(null);
-    }
-
-    const fetchN1 = async () => {
-      diagrams.setN1Loading(true);
-      // On user-initiated contingency change, switch to N-1 so the
-      // fetched diagram is visible. On session restore, leave the
-      // active tab alone so the user lands on whichever tab the
-      // VisualizationPanel default (Action / Overflow) resolves to
-      // given the restored state.
-      if (!isRestoring) {
-        diagrams.setActiveTab('n-1');
-      }
-
-      // Fast path — svgPatch DOM-recycling. Clone the N-state SVG and
-      // mutate only what changed (the dashed contingency line, flow
-      // labels, overload halos) instead of fetching a fresh ~20 MB
-      // NAD. Falls back to the full /api/n1-diagram endpoint on any
-      // error or when the base N DOM is not yet mounted. Session
-      // reload always uses the full-fetch path to preserve the
-      // reload contract. See docs/performance/history/svg-dom-recycling.md.
-      const baseSvgEl = diagrams.nSvgContainerRef.current?.querySelector('svg') as SVGSVGElement | null;
-      const baseMeta = diagrams.nMetaIndex;
-      const canPatch = !!baseSvgEl && !!baseMeta && !isRestoring;
-      if (canPatch) {
-        console.log('[svgPatch] N-1 patch PATH — calling /api/n1-diagram-patch', selectedBranch);
-        try {
-          const patch = await api.getN1DiagramPatch(selectedBranch);
-          if (patch.patchable) {
-            console.log('[svgPatch] N-1 patch applied (patchable=true)', selectedBranch);
-            const cloned = cloneBaseSvg(baseSvgEl!);
-            applyPatchToClone(cloned, baseMeta!, patch);
-            diagrams.setN1Diagram({
-              svg: cloned,
-              metadata: nDiagram?.metadata ?? null,
-              originalViewBox: nDiagram?.originalViewBox ? { ...nDiagram.originalViewBox } : null,
-              lines_overloaded: patch.lines_overloaded,
-              lines_overloaded_rho: patch.lines_overloaded_rho,
-              flow_deltas: patch.flow_deltas,
-              reactive_flow_deltas: patch.reactive_flow_deltas,
-              asset_deltas: patch.asset_deltas,
-              lf_converged: patch.lf_converged,
-              lf_status: patch.lf_status,
-            });
-            diagrams.lastZoomState.current = { query: '', branch: '' };
-            diagrams.setN1Loading(false);
-            return;
-          }
-        } catch (e) {
-          console.warn('[svgPatch] N-1 patch threw — falling back to full fetch', e);
-        }
-      } else {
-        console.log('[svgPatch] N-1 patch SKIPPED — no base N SVG/meta yet. baseSvgEl:', !!baseSvgEl, 'baseMeta:', !!baseMeta, 'restoring:', isRestoring);
-      }
-
-      try {
-        const res = await api.getN1Diagram(selectedBranch);
-        const { svg, viewBox } = processSvg(res.svg, voltageLevels.length);
-        diagrams.setN1Diagram({ ...res, svg, originalViewBox: viewBox });
-        // Ensure auto-zoom fires after the new N-1 diagram is ready.
-        // Reset lastZoomState so the auto-zoom effect sees a "branch change"
-        // on the render that has both activeTab='n-1' and the new SVG in DOM.
-        diagrams.lastZoomState.current = { query: '', branch: '' };
-      } catch (err) {
-        console.error('Failed to fetch N-1 diagram', err);
-        setError(`Failed to fetch N-1 diagram for ${selectedBranch}`);
-      } finally {
-        diagrams.setN1Loading(false);
-      }
-    };
-    fetchN1();
-  }, [selectedBranch, branches, voltageLevels.length, hasAnalysisState, clearContingencyState, analysisLoading, n1Diagram, n1Loading, setError, diagrams, nDiagram]);
+  useN1Fetch({
+    selectedBranch,
+    branches,
+    voltageLevelsLength: voltageLevels.length,
+    diagrams,
+    analysisLoading,
+    hasAnalysisState,
+    clearContingencyState,
+    setSelectedBranch,
+    setConfirmDialog,
+    setError,
+  });
 
   useEffect(() => {
     const nextSet = n1Diagram?.lines_overloaded ? new Set(n1Diagram.lines_overloaded) : new Set<string>();
@@ -1003,208 +904,14 @@ function App() {
 
 
 
-  const staleHighlights = useRef<Set<TabId>>(new Set());
-  const prevHighlightTabRef = useRef<TabId>(activeTab);
-
-  // Flow vs Impacts view mode is now TAB-SCOPED and WINDOW-SCOPED:
-  // main window has `actionViewMode` (from useDiagrams) while each
-  // detached tab has its own entry here. Toggling Impacts in a
-  // detached popup therefore only affects that popup's tab; the
-  // main window's mode is untouched, and vice versa. A reattach
-  // clears the detached-mode entry so the tab resumes the
-  // main-window mode from that point onward.
-  const [detachedViewModes, setDetachedViewModes] = useState<Partial<Record<TabId, 'network' | 'delta'>>>({});
-
-  // Which Flow/Impacts mode applies to a given tab right now.
-  const viewModeForTab = useCallback((tab: TabId): 'network' | 'delta' => {
-    if (detachedTabs[tab]) return detachedViewModes[tab] ?? 'network';
-    return actionViewMode;
-  }, [detachedTabs, detachedViewModes, actionViewMode]);
-
-  // Per-tab Flow/Impacts toggle routing: if the tab is currently
-  // detached, update its entry in `detachedViewModes`; otherwise
-  // update the main-window `actionViewMode`. This is how a Flow
-  // /Impacts button inside a detached popup affects ONLY the
-  // detached window.
-  const handleViewModeChangeForTab = useCallback((tab: TabId, mode: 'network' | 'delta') => {
-    interactionLogger.record('view_mode_changed', { mode, tab, scope: detachedTabs[tab] ? 'detached' : 'main' });
-    if (detachedTabs[tab]) {
-      setDetachedViewModes(prev => ({ ...prev, [tab]: mode }));
-    } else {
-      diagrams.setActionViewMode(mode);
-    }
-  }, [detachedTabs, diagrams]);
-
-  // On reattach, drop the tab's detached view-mode entry so the
-  // main window's `actionViewMode` takes over for that tab from
-  // that point onward.
-  useEffect(() => {
-    setDetachedViewModes(prev => {
-      const next: Partial<Record<TabId, 'network' | 'delta'>> = {};
-      let changed = false;
-      for (const tabId of Object.keys(prev) as TabId[]) {
-        if (detachedTabs[tabId]) {
-          next[tabId] = prev[tabId];
-        } else {
-          changed = true; // dropped
-        }
-      }
-      return changed ? next : prev;
-    });
-  }, [detachedTabs]);
-
-  const applyHighlightsForTab = useCallback((tab: TabId, mode?: 'network' | 'delta') => {
-    const effectiveMode = mode ?? viewModeForTab(tab);
-    const overloadedLines = result?.lines_overloaded || [];
-
-    // For the N-1 tab, the overloads are known as soon as the N-1
-    // diagram comes back from the backend (`n1Diagram.lines_overloaded`)
-    // — well before any action analysis has run. Falling back to that
-    // list lets the orange overload halos show up immediately on the
-    // N-1 view, matching the standalone interface and what the user
-    // expects to see right after picking a contingency. Once analysis
-    // runs and `result.lines_overloaded` becomes available, we use that.
-    // In both cases the user's selected-overload set (from the
-    // Overloads panel) further filters the list down. Extracted to a
-    // pure helper for unit testing.
-    const n1OverloadedLines = computeN1OverloadHighlights(
-      overloadedLines,
-      n1Diagram?.lines_overloaded,
-      selectedOverloads,
-    );
-
-    if (tab === 'n-1') {
-      if (diagrams.n1SvgContainerRef.current) {
-        // IMPORTANT: run highlight CLONES before applyDeltaVisuals.
-        // The clone-based highlights (`applyOverloadedHighlights`,
-        // `applyContingencyHighlight`) use cloneNode(true) on the
-        // original SVG element. If applyDeltaVisuals has already
-        // tagged the original with `nad-delta-positive/negative/grey`,
-        // the clone inherits that class — and because the `.nad-delta-*`
-        // CSS is declared LATER in App.css than `.nad-contingency-highlight`
-        // / `.nad-overloaded`, the delta rule wins the cascade and the
-        // halo becomes a thin orange/blue stroke, effectively making the
-        // highlight disappear in Impacts mode. Cloning first guarantees
-        // the halos stay on a pristine copy of the element.
-        //
-        // Overloaded lines must also be highlighted in BOTH Flows and
-        // Impacts modes — the user looks at the Impacts view to see how
-        // the action redistributes flows AND which lines are still
-        // (or newly) overloaded; suppressing the halos in delta mode
-        // hides exactly that information.
-        if (diagrams.n1MetaIndex && n1OverloadedLines.length > 0) {
-          applyOverloadedHighlights(diagrams.n1SvgContainerRef.current, diagrams.n1MetaIndex, n1OverloadedLines);
-        }
-        applyContingencyHighlight(diagrams.n1SvgContainerRef.current, diagrams.n1MetaIndex, selectedBranch);
-        applyDeltaVisuals(diagrams.n1SvgContainerRef.current, n1Diagram, diagrams.n1MetaIndex, effectiveMode === 'delta');
-      }
-    }
-    if (tab === 'action') {
-      const actionDetail = result?.actions?.[selectedActionId || ''];
-
-      if (actionDetail) {
-        // Same ordering rule as the N-1 tab: clone-based highlights
-        // first (so they capture pristine elements), delta visuals
-        // last. Overload halos render in both Flows and Impacts modes.
-        //
-        // ActionCard severity classification (`components/ActionCard.tsx:76-78`):
-        //   - red:    `max_rho > monitoringFactor`         ("Still overloaded")
-        //   - orange: `max_rho > monitoringFactor - 0.05`  ("Solved — low margin")
-        //   - green:  else                                 ("Solves overload")
-        // When the card says "Solved" (orange or green) no line in the
-        // post-action state exceeds the operator's tolerance threshold,
-        // so overload halos on the NAD contradict the card's label.
-        // The backend's manual-simulation path flags `lines_overloaded_after`
-        // on raw rho >= monitoring_factor (0.95) whereas the analysis
-        // path uses raw rho >= 1.0 (`analysis_mixin.py:97` vs
-        // `simulation_mixin.py:536`). In the low-margin band — raw rho
-        // in `[mf, 1.0)` — the two paths disagree: manual ships a
-        // non-empty list, analysis ships an empty one. Without this
-        // gate, halos appear on manual low-margin actions but not on
-        // suggested ones with the same displayed max_rho — the exact
-        // user-reported bug (commit-time 2026-04-20).
-        const isSolved =
-          actionDetail.max_rho != null && actionDetail.max_rho <= monitoringFactor;
-
-        let overloadsToHighlight: string[] = [];
-
-        if (!isSolved) {
-          if (actionDetail.lines_overloaded_after && actionDetail.lines_overloaded_after.length > 0) {
-            overloadsToHighlight = actionDetail.lines_overloaded_after;
-          } else {
-            // Fallback for legacy results or actions without full enrichment
-            if (overloadedLines.length > 0 && actionDetail.rho_after) {
-              overloadedLines.forEach((name, i) => {
-                const rho = actionDetail.rho_after![i];
-                if (rho != null && rho > monitoringFactor) {
-                  overloadsToHighlight.push(name);
-                }
-              });
-            }
-            if (actionDetail.max_rho != null && actionDetail.max_rho > monitoringFactor && actionDetail.max_rho_line) {
-              if (!overloadsToHighlight.includes(actionDetail.max_rho_line)) {
-                overloadsToHighlight.push(actionDetail.max_rho_line);
-              }
-            }
-          }
-        }
-
-        if (diagrams.actionSvgContainerRef.current && diagrams.actionMetaIndex) {
-          applyOverloadedHighlights(diagrams.actionSvgContainerRef.current, diagrams.actionMetaIndex, overloadsToHighlight);
-        }
-
-        if (diagrams.actionSvgContainerRef.current) {
-          applyActionTargetHighlights(diagrams.actionSvgContainerRef.current, diagrams.actionMetaIndex, actionDetail, selectedActionId);
-          applyContingencyHighlight(diagrams.actionSvgContainerRef.current, diagrams.actionMetaIndex, selectedBranch);
-        }
-      }
-      else {
-        if (diagrams.actionSvgContainerRef.current) {
-          applyActionTargetHighlights(diagrams.actionSvgContainerRef.current, null, null, null);
-        }
-      }
-
-      // Delta visuals run LAST so they decorate the originals without
-      // contaminating the highlight clones already in the background
-      // layer.
-      applyDeltaVisuals(diagrams.actionSvgContainerRef.current, actionDiagram, diagrams.actionMetaIndex, effectiveMode === 'delta');
-    }
-  }, [n1Diagram, actionDiagram, result, selectedActionId, selectedBranch, diagrams, monitoringFactor, viewModeForTab, selectedOverloads]);
-
-  useEffect(() => {
-    const isTabSwitch = prevHighlightTabRef.current !== activeTab;
-    prevHighlightTabRef.current = activeTab;
-    const otherTabs: TabId[] = ['n', 'n-1', 'action'].filter(t => t !== activeTab) as TabId[];
-    otherTabs.forEach(t => staleHighlights.current.add(t));
-
-    // Apply highlights + delta visuals to both the main window's
-    // active tab AND every currently-detached tab — because Impacts
-    // mode must keep working inside a detached popup when only the
-    // popup's view mode changes (the main window may be showing a
-    // different tab entirely).
-    const applyAll = () => {
-      applyHighlightsForTab(activeTab);
-      staleHighlights.current.delete(activeTab);
-      for (const detachedId of Object.keys(detachedTabs) as TabId[]) {
-        if (detachedId === activeTab) continue;
-        if (detachedId === 'overflow') continue;
-        applyHighlightsForTab(detachedId);
-        staleHighlights.current.delete(detachedId);
-      }
-    };
-
-    if (isTabSwitch) {
-      // Double rAF to ensure browser layout is settled before getScreenCTM()
-      const id = requestAnimationFrame(() => {
-        requestAnimationFrame(() => {
-          applyAll();
-        });
-      });
-      return () => cancelAnimationFrame(id);
-    } else {
-      applyAll();
-    }
-  }, [nDiagram, n1Diagram, actionDiagram, diagrams.nMetaIndex, diagrams.n1MetaIndex, diagrams.actionMetaIndex, result, selectedActionId, actionViewMode, detachedViewModes, detachedTabs, activeTab, selectedBranch, selectedOverloads, applyHighlightsForTab]);
+  const { viewModeForTab, handleViewModeChangeForTab } = useDiagramHighlights({
+    diagrams,
+    result,
+    selectedBranch,
+    selectedOverloads,
+    monitoringFactor,
+    detachedTabs,
+  });
 
   // ===== Extracted JSX callbacks (stable references for React.memo) =====
 
@@ -1302,179 +1009,77 @@ function App() {
       />
 
       <div style={{ flex: 1, display: 'flex', overflow: 'hidden' }}>
-        {/*
-          Sidebar layout:
-          - A COMPACT sticky strip at the top keeps only the
-            clickable fields of interest visible while scrolling
-            (selected contingency → zoom active tab; N-1 overloads →
-            jump to N-1 tab + zoom, same behavior as the old
-            "Loading Before" link on action cards).
-          - Everything else — the full Select Contingency card with
-            the search input, the Overloads panel with its warnings
-            and N/N-1 breakdown, and the ActionFeed — scrolls
-            together in a single column below, saving vertical space.
-        */}
-        <div data-testid="sidebar" style={{ width: '25%', background: '#eee', borderRight: '1px solid #ccc', display: 'flex', flexDirection: 'column', overflow: 'hidden' }}>
-          {(selectedBranch || (n1Diagram?.lines_overloaded?.length ?? 0) > 0) && (
-            <div
-              data-testid="sticky-feed-summary"
-              style={{
-                flexShrink: 0,
-                padding: '6px 12px',
-                background: '#f8f9fa',
-                borderBottom: '1px solid #ccc',
-                fontSize: '11px',
-                lineHeight: 1.5,
-                boxShadow: '0 1px 2px rgba(0,0,0,0.04)',
-              }}
-            >
-              {selectedBranch && (
-                <div style={{ display: 'flex', alignItems: 'baseline', gap: '4px' }}>
-                  <span style={{ color: '#555', fontWeight: 600, whiteSpace: 'nowrap' }}>🎯 Contingency:</span>
-                  <button
-                    onClick={(e) => { e.stopPropagation(); handleZoomOnActiveTab(selectedBranch); }}
-                    title={`Zoom to ${selectedBranch} in the current diagram`}
-                    style={{
-                      background: 'none',
-                      border: 'none',
-                      cursor: 'pointer',
-                      padding: 0,
-                      fontSize: '11px',
-                      color: '#1e40af',
-                      fontWeight: 600,
-                      textDecoration: 'underline dotted',
-                      wordBreak: 'break-word',
-                      textAlign: 'left',
-                    }}
-                  >
-                    🔍 {displayName(selectedBranch)}
-                  </button>
-                </div>
-              )}
-              {(n1Diagram?.lines_overloaded?.length ?? 0) > 0 && (
-                <div style={{ display: 'flex', alignItems: 'baseline', gap: '4px' }}>
-                  <span style={{ color: '#b91c1c', fontWeight: 600, whiteSpace: 'nowrap' }}>⚠️ N-1:</span>
-                  <span style={{ wordBreak: 'break-word' }}>
-                    {n1Diagram!.lines_overloaded!.map((name, i) => {
-                      const rho = n1Diagram!.lines_overloaded_rho?.[i];
-                      const rhoPct = rho != null && !Number.isNaN(rho) ? `${(rho * 100).toFixed(1)}%` : null;
-                      const isSelected = selectedOverloads?.has(name) ?? true;
-                      return (
-                        <React.Fragment key={name}>
-                          {i > 0 && ', '}
-                          <button
-                            onClick={(e) => { e.stopPropagation(); wrappedAssetClick('', name, 'n-1'); }}
-                            title={`Open N-1 tab and zoom to ${name}`}
-                            style={{
-                              background: 'none',
-                              border: 'none',
-                              cursor: 'pointer',
-                              padding: 0,
-                              fontSize: '11px',
-                              color: isSelected ? '#1e40af' : '#bdc3c7',
-                              fontWeight: isSelected ? 600 : 400,
-                              textDecoration: isSelected ? 'underline dotted' : 'none',
-                            }}
-                          >
-                            {displayName(name)}
-                          </button>
-                          {rhoPct && (
-                            <span style={{ color: isSelected ? '#374151' : '#bdc3c7', marginLeft: '2px' }}>
-                              ({rhoPct})
-                            </span>
-                          )}
-                        </React.Fragment>
-                      );
-                    })}
-                  </span>
-                </div>
-              )}
-            </div>
-          )}
-          <div style={{ flex: 1, overflowY: 'auto', padding: '15px', minHeight: 0, display: 'flex', flexDirection: 'column', gap: '15px' }}>
-            {/* Target Contingency selector — full card lives in the
-                scroll area so the compact sticky strip above stays
-                minimal. */}
-            {branches.length > 0 && (
-              <div style={{ flexShrink: 0, padding: '10px 15px', background: 'white', borderRadius: '8px', border: '1px solid #dee2e6', boxShadow: '0 2px 4px rgba(0,0,0,0.05)' }}>
-                <label style={{ fontSize: '0.8rem', fontWeight: 'bold', display: 'block', marginBottom: '5px' }}>🎯 Select Contingency</label>
-                <input
-                  list="contingencies"
-                  value={selectedBranch}
-                  onChange={handleContingencyChange}
-                  placeholder="Search line/bus..."
-                  style={{ width: '100%', padding: '7px 10px', border: '1px solid #ccc', borderRadius: '4px', boxSizing: 'border-box', fontSize: '0.85rem' }}
-                />
-                {selectedBranch && nameMap[selectedBranch] && (
-                  <div style={{ fontSize: '0.78rem', color: '#4b5563', marginTop: '3px', fontStyle: 'italic', lineHeight: 1.3 }}>
-                    {nameMap[selectedBranch]}
-                  </div>
-                )}
-                <datalist id="contingencies">
-                  {contingencyOptions}
-                </datalist>
-              </div>
-            )}
-
-            <div style={{ flexShrink: 0 }}>
-              <OverloadPanel
-                nOverloads={nDiagram?.lines_overloaded || []}
-                n1Overloads={n1Diagram?.lines_overloaded || []}
-                nOverloadsRho={nDiagram?.lines_overloaded_rho}
-                n1OverloadsRho={n1Diagram?.lines_overloaded_rho}
-                onAssetClick={wrappedAssetClick as (actionId: string, assetName: string, tab?: 'n' | 'n-1') => void}
-                showMonitoringWarning={showMonitoringWarning}
-                monitoredLinesCount={monitoredLinesCount}
-                totalLinesCount={totalLinesCount}
-                monitoringFactor={monitoringFactor}
-                preExistingOverloadThreshold={preExistingOverloadThreshold}
-                onDismissWarning={handleDismissWarning}
-                onOpenSettings={handleOpenConfigSettings}
-                selectedOverloads={selectedOverloads}
-                onToggleOverload={analysis.handleToggleOverload}
-                monitorDeselected={monitorDeselected}
-                onToggleMonitorDeselected={handleToggleMonitorDeselected}
-                displayName={displayName}
-              />
-            </div>
-            <ActionFeed
-              actions={result?.actions || {}}
-              actionScores={result?.action_scores}
-              linesOverloaded={result?.lines_overloaded || []}
-              selectedActionId={selectedActionId}
-              scrollTarget={scrollTarget}
-              selectedActionIds={selectedActionIds}
-              rejectedActionIds={rejectedActionIds}
-              manuallyAddedIds={manuallyAddedIds}
-              combinedActions={result?.combined_actions ?? null}
-              pendingAnalysisResult={pendingAnalysisResult}
-              onDisplayPrioritizedActions={wrappedDisplayPrioritized}
-              onRunAnalysis={wrappedRunAnalysis}
-              canRunAnalysis={!!selectedBranch && !analysisLoading}
-              onActionSelect={wrappedActionSelect}
-              onActionFavorite={wrappedActionFavorite}
-              onActionReject={actionsHook.handleActionReject}
-              onAssetClick={wrappedAssetClick}
-              nodesByEquipmentId={diagrams.nMetaIndex?.nodesByEquipmentId ?? null}
-              edgesByEquipmentId={diagrams.nMetaIndex?.edgesByEquipmentId ?? null}
-              disconnectedElement={selectedBranch || null}
-              onManualActionAdded={wrappedManualActionAdded}
-              onActionResimulated={wrappedActionResimulated}
-              analysisLoading={analysisLoading}
+        <AppSidebar
+          selectedBranch={selectedBranch}
+          branches={branches}
+          nameMap={nameMap}
+          n1LinesOverloaded={n1Diagram?.lines_overloaded}
+          n1LinesOverloadedRho={n1Diagram?.lines_overloaded_rho}
+          selectedOverloads={selectedOverloads}
+          contingencyOptions={contingencyOptions}
+          onContingencyChange={handleContingencyChange}
+          displayName={displayName}
+          onContingencyZoom={handleZoomOnActiveTab}
+          onOverloadClick={wrappedAssetClick as (actionId: string, assetName: string, tab: 'n' | 'n-1') => void}
+        >
+          <div style={{ flexShrink: 0 }}>
+            <OverloadPanel
+              nOverloads={nDiagram?.lines_overloaded || []}
+              n1Overloads={n1Diagram?.lines_overloaded || []}
+              nOverloadsRho={nDiagram?.lines_overloaded_rho}
+              n1OverloadsRho={n1Diagram?.lines_overloaded_rho}
+              onAssetClick={wrappedAssetClick as (actionId: string, assetName: string, tab?: 'n' | 'n-1') => void}
+              showMonitoringWarning={showMonitoringWarning}
+              monitoredLinesCount={monitoredLinesCount}
+              totalLinesCount={totalLinesCount}
               monitoringFactor={monitoringFactor}
-              onVlDoubleClick={handleVlDoubleClick}
-              recommenderConfig={recommenderConfig}
-              actionDictFileName={actionDictFileName}
-              actionDictStats={actionDictStats}
-              onOpenSettings={handleOpenSettings}
-              onUpdateCombinedEstimation={handleUpdateCombinedEstimation}
+              preExistingOverloadThreshold={preExistingOverloadThreshold}
+              onDismissWarning={handleDismissWarning}
+              onOpenSettings={handleOpenConfigSettings}
+              selectedOverloads={selectedOverloads}
+              onToggleOverload={analysis.handleToggleOverload}
+              monitorDeselected={monitorDeselected}
+              onToggleMonitorDeselected={handleToggleMonitorDeselected}
               displayName={displayName}
-              onActionDiagramPrimed={diagrams.primeActionDiagram}
-              voltageLevelsLength={voltageLevels.length}
-              overviewFilters={overviewFilters}
             />
           </div>
-        </div>
+          <ActionFeed
+            actions={result?.actions || {}}
+            actionScores={result?.action_scores}
+            linesOverloaded={result?.lines_overloaded || []}
+            selectedActionId={selectedActionId}
+            scrollTarget={scrollTarget}
+            selectedActionIds={selectedActionIds}
+            rejectedActionIds={rejectedActionIds}
+            manuallyAddedIds={manuallyAddedIds}
+            combinedActions={result?.combined_actions ?? null}
+            pendingAnalysisResult={pendingAnalysisResult}
+            onDisplayPrioritizedActions={wrappedDisplayPrioritized}
+            onRunAnalysis={wrappedRunAnalysis}
+            canRunAnalysis={!!selectedBranch && !analysisLoading}
+            onActionSelect={wrappedActionSelect}
+            onActionFavorite={wrappedActionFavorite}
+            onActionReject={actionsHook.handleActionReject}
+            onAssetClick={wrappedAssetClick}
+            nodesByEquipmentId={diagrams.nMetaIndex?.nodesByEquipmentId ?? null}
+            edgesByEquipmentId={diagrams.nMetaIndex?.edgesByEquipmentId ?? null}
+            disconnectedElement={selectedBranch || null}
+            onManualActionAdded={wrappedManualActionAdded}
+            onActionResimulated={wrappedActionResimulated}
+            analysisLoading={analysisLoading}
+            monitoringFactor={monitoringFactor}
+            onVlDoubleClick={handleVlDoubleClick}
+            recommenderConfig={recommenderConfig}
+            actionDictFileName={actionDictFileName}
+            actionDictStats={actionDictStats}
+            onOpenSettings={handleOpenSettings}
+            onUpdateCombinedEstimation={handleUpdateCombinedEstimation}
+            displayName={displayName}
+            onActionDiagramPrimed={diagrams.primeActionDiagram}
+            voltageLevelsLength={voltageLevels.length}
+            overviewFilters={overviewFilters}
+          />
+        </AppSidebar>
         <div style={{ flex: 1, background: 'white', display: 'flex', flexDirection: 'column' }}>
           <VisualizationPanel
             activeTab={activeTab}
@@ -1545,29 +1150,7 @@ function App() {
         onCancel={handleCancelDialog}
         onConfirm={handleConfirmDialog}
       />
-      {error && (
-        <div style={{
-          position: 'fixed', bottom: 20, right: 20,
-          background: '#e74c3c', color: 'white',
-          padding: '10px 20px', borderRadius: '4px',
-          boxShadow: '0 2px 10px rgba(0,0,0,0.2)', zIndex: 1000,
-        }}>
-          {error}
-        </div>
-      )}
-      {infoMessage && (
-        <div style={{
-          position: 'fixed', bottom: 20, left: 20,
-          background: infoMessage.startsWith('SUCCESS') ? '#27ae60' : '#3498db',
-          color: 'white',
-          padding: '12px 24px', borderRadius: '4px',
-          boxShadow: '0 4px 15px rgba(0,0,0,0.3)', zIndex: 1000,
-          fontWeight: 'bold',
-          border: '1px solid rgba(255,255,255,0.2)'
-        }}>
-          {infoMessage}
-        </div>
-      )}
+      <StatusToasts error={error} infoMessage={infoMessage} />
     </div>
   );
 }
